float volume value

+++ examples/dummy_daemon.py
@@ -13,7 +13,7 @@ class MyPlugin(PluginInterface):
         return self.muted
         
     def update_volume(self, volume):
-        print("[Plugin] volume changed to %d"%volume)
+        print("[Plugin] volume changed to %.1f"%volume)
         self.volume = volume
         
     def update_muted(self, muted):
@@ -24,9 +24,9 @@ class MyPlugin(PluginInterface):
 if __name__ == "__main__":
     plugin = MyPlugin()
     eh = EventHandler(plugin)
-    print("Current volume: %d"%plugin.volume)
+    print("Current volume: %.1f"%plugin.volume)
     while True:
         newvol = input("Enter new volume: ")
-        plugin.volume = int(newvol)
+        plugin.volume = float(newvol)
         eh.on_plugin_change()
